Fixes #55 - Simplifies fixture creation and moves a testing class to the package level
